fix #9: 'text' type requests should not return null
The 'text' type request returns the last known 'status' (not 'show'). It's a bit dodgy that this takes the last known status when a user is offline (where other endpoints return a 'current' status). This behavior is not changed in this commit (warrants further investigation - it is cleary implemented intentionally).

When the status is absent, this endpoint returns a value `null`. The status is an optional element, not set by all clients. It is undesirable to have a 'null' value. This commit replaces it with an hardcoded value of 'Available', matching the implementation that hardcodedly sets 'Unavailable' for clients that are offline _and_ do not have a 'last known' presence.